Changed Tweet ids over to BigInts to account for 64-bit twitter ids, fixed setting last_seen_id, fully changed check_user to GET request

+++ apps/account/models.py
@@ -24,7 +24,7 @@ class User(models.Model):
 class Follow(models.Model):
     user = models.ForeignKey(User)
     screen_name = models.CharField(max_length=20)
-    last_id_seen = models.IntegerField(blank=True, null=True)
+    last_id_seen = models.BigIntegerField(blank=True, null=True)
     timestamp = models.DateTimeField(auto_now_add=True)
 
     class Meta:
